fix(bed-data-fetcher): change fetcher name
