+++ src/data-fetchers/index.ts
@@ -3,4 +3,4 @@ export { default as VcfDataFetcher } from './vcf/vcf-data-fetcher';
 export { default as BigWigDataFetcher } from './bigwig/bigwig-data-fetcher';
 export { default as CsvDataFetcher } from './csv/csv-data-fetcher';
 export { default as JsonDataFetcher } from './json/json-data-fetcher';
-export { default as BedDataFetcher } from './bed/bed-data-fetcher-tabix';
+export { default as BedDataFetcher } from './bed/bed-data-fetcher';
